requirements.txt
PyYAML
numpy
pandas
scikit-learn
torch
transformers
pytz
requests
aiohttp
matplotlib
seaborn
ta
optuna
textblob
tensorflow
yfinance
forex-python
tf-keras
aioschedule
backoff
yahoo_fin
tenacity
yahooquery
MetaTrader5